% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird_pars_wide_example}
\alias{bird_pars_wide_example}
\title{Example of bird parameters stored in wide format}
\format{
A 3 x 16 data frame, with the biological parameters (columns) for
each of the 3 species (rows). Columns include:
\describe{
\item{Species}{Species name}
\item{AvoidanceBasic}{Mean of basic avoidance rate}
\item{AvoidanceBasicSD}{SD of basic avoidance rate}
\item{AvoidanceExtended}{Mean of extended avoidance rate}
\item{AvoidanceExtendedSD}{SD of extended avoidance rate}
\item{Body_Length}{Mean body length}
\item{Body_LengthSD}{SD of body length}
...
}
}
\usage{
bird_pars_wide_example
}
\description{
A data frame of (fake) data on biological parameters of three seabird
species.
}
\details{
Intended to illustrate the application of \code{stoch_scrm()} to a multiple
scenario setting, where parameter data is available from tables in wide format.
}
\keyword{datasets}
