% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Count_Data}
\alias{Count_Data}
\title{Bird densities for the CRM}
\format{
A dataframe
\describe{
\item{Species}{Species name}
\item{AvoidanceBasic}{Basic avoidance rate}
\item{AvoidanceBasicSD}{Basic avoidance rate SD}
...
}
}
\usage{
Count_Data
}
\description{
A dataframe of Bird densities for each month
}
\keyword{datasets}
