% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CountData}
\alias{CountData}
\title{Bird densities for the CRM}
\format{
A dataframe
\describe{
  \item{Species}{Species name}
  \item{AvoidanceBasic}{Basic avoidance rate}
  \item{AvoidanceBasicSD}{Basic avoidance rate SD}
  ...
}
}
\usage{
CountData
}
\description{
A dataframe of Bird densities for each month
}
\keyword{datasets}
