% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverC}
\alias{coverC}
\title{Rotor blade sizes}
\format{
A dataframe
\describe{
  \item{rad}{radius}
  \item{circ}{Circumference}
  ...
}
}
\usage{
coverC
}
\description{
A dataframe giving the width of the chord relative to its maximum width at given points along the radius of the rotor
}
\keyword{datasets}
