% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_opt2.R
\name{crm_opt2}
\alias{crm_opt2}
\title{Number of collisions under model option 2}
\usage{
crm_opt2(
  gen_d_y,
  flux_factor,
  prob_single_collision,
  prop_operational,
  avoidance_rate,
  lac_factor
)
}
\arguments{
\item{gen_d_y}{a vector with the proportion of birds at height bands across #
the rotor disc, calculated from a generic flight height distribution}

\item{flux_factor}{a vector containing the flux factor for each month}

\item{prob_single_collision}{The probability of collision for a single bird
transit through a rotor, assuming no avoidance action.}

\item{prop_operational}{The proportion of time during which turbines are
operational per month}

\item{avoidance_rate}{The proportion of birds likely to take action to avoid
collision with rotor}

\item{lac_factor}{The large array correction factor. Defaults to 1,
meaning large array correction is not applicable}
}
\value{
Expected number of collisions per month based on model option 2
}
\description{
Top-level function to run CRM model under option 2:
   \itemize{
      \item Basic model, i.e. flights across collision risk height are
       uniformly distributed
      \item Proportion at collision risk height derived from a generic flight
      height distribution
   }
}
\seealso{
[get_fhd_rotor()], [get_flux_factor()]
}
