% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_opt3.R
\name{crm_opt3}
\alias{crm_opt3}
\title{Number of collisions under model option 3}
\usage{
crm_opt3(
  gen_d_y,
  rotor_radius,
  blade_width,
  rotor_speed,
  blade_pitch,
  flight_type_num,
  wing_span,
  flight_speed,
  body_lt,
  n_blades,
  prop_upwind,
  avoidance_rate_ext,
  flux_factor,
  prop_operational,
  lac_factor
)
}
\arguments{
\item{gen_d_y}{a vector with the proportion of birds at height bands across #
the rotor disc, calculated from a generic flight height distribution}

\item{flux_factor}{a vector containing the flux factor for each month}

\item{prop_operational}{The proportion of time during which turbines are
operational per month}

\item{lac_factor}{The large array correction factor. Defaults to 1,
meaning large array correction is not applicable}
}
\value{
Expected number of collisions per month based on model option 2
}
\description{
Top-level function to run CRM model under option 3:
   \itemize{
      \item Extended model, i.e. takes into account the distribution of bird
      flight heights at risk height (above the minimum and below the maximum
      height of the rotor blade)
      \item Uses a generic flight height distribution
   }
}
\note{
The original Masden code included the calculation of the "Flux Integral" and
   "Average collision risk for single rotor transit". These quantities are not
   required for the calculation of collisions under option 3, nor are used
   anywhere else in the remaining code. Therefore, those calculations were
   dropped from the current implementation.
}
\seealso{
[get_fhd_rotor()], [get_flux_factor()]
}
