% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CV}
\alias{CV}
\title{Coefficient of Variation}
\usage{
CV(mean, sd)
}
\arguments{
\item{mean}{numeric, mean of the random variable}

\item{sd}{numeric, standard deviation of the random variable}
}
\value{
a numeric value
}
\description{
Calculate the coefficient of variation of a single value
}
