% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayLength.r
\name{DayLength}
\alias{DayLength}
\title{Total day and night hours per month}
\usage{
DayLength(wf_latitude)
}
\arguments{
\item{wf_latitude}{A decimal value. The latitude of the centroid of the
windfarm, in degrees.}
}
\value{
data frame with total number of daylight hours and night hours in each
month at the specified latitude.
}
\description{
Taken from Forsythe et al.(1995) A model comparison for daylength as a
function of latitude and day of year.  Ecological Modelling. 80: 87 - 95
}
\examples{
    x <- DayLength(54.6)

}
