% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstup}
\alias{firstup}
\title{Upper-case conversion of the first character of a string}
\usage{
firstup(x)
}
\arguments{
\item{x}{a character string}
}
\description{
Upper-case conversion of the first character of a string
}
