% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_months}
\alias{format_months}
\title{Format month names}
\usage{
format_months(months)
}
\arguments{
\item{months}{a character vector with month names}
}
\description{
Formats a character vector with month names to their 3-letter abbreviation
}
\examples{
   format_months(c("Jan","Feb"))
}
