% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collisions_basic.R
\name{get_collisions_basic}
\alias{get_collisions_basic}
\title{Get expected collisions based on the basic model}
\usage{
get_collisions_basic(
  n_transits,
  prob_single_collision,
  prop_operational,
  avoidance_rate,
  lac_factor = 1
)
}
\arguments{
\item{n_transits}{Vector with the estimated number of bird flights crossing
the rotors of the wind farm at each time period}

\item{prob_single_collision}{The probability of collision for a single bird
transit through a rotor, assuming no avoidance action.}

\item{prop_operational}{The proportion of time during which turbines are
operational per month}

\item{avoidance_rate}{The proportion of birds likely to take action to avoid
collision with rotor}

\item{lac_factor}{The large array correction factor. Defaults to 1,
meaning large array correction is not applicable}
}
\value{
A numeric vector. The expected number of collisions at each time period
}
\description{
Provides the number of collisions expected to occur by month. The basic model
assumes a uniform distribution of bird flights at risk height (i.e. between
min and max rotor height)
}
