% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lac_factor.R
\name{get_lac_factor}
\alias{get_lac_factor}
\title{Large array correction factor}
\usage{
get_lac_factor(
  n_turbines,
  rotor_radius,
  avoidance_rate,
  prob_single_collision,
  mean_prop_operational,
  wf_width
)
}
\arguments{
\item{n_turbines}{The number of turbines on the wind farm (\eqn{T}).}

\item{rotor_radius}{The radius of the rotor (\eqn{R}), in metres}

\item{avoidance_rate}{The avoidance rate assumed (\eqn{A})}

\item{prob_single_collision}{The average probability of collision for a single bird
transit through a rotor, assuming no avoidance action.}

\item{mean_prop_operational}{the mean proportion of time turbines are
operational through the year (decimal)}

\item{wf_width}{The approximate longitudinal width of the wind farm,
in kilometers (\eqn{w})}
}
\value{
The large array correction factor to be applied
}
\description{
A correction factor for large windfarms with a large array of turbines, to take
into account the depletion of bird density in later rows of the site.
}
