% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_months.R
\name{get_months}
\alias{get_months}
\title{Gets a list of months from a month range}
\usage{
get_months(Mrange, month.abb)
}
\arguments{
\item{Mrange}{A character string. The month range in the format "XXX - XXX"}
}
\value{
A vector of the months within the given range
}
\description{
Outputs a vector of months from the input month range
}
\examples{
get_months("Jan - Apr")

}
