% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_collision.R
\name{get_prob_collision}
\alias{get_prob_collision}
\title{Single transit collision risk with no avoidance}
\usage{
get_prob_collision(
  chord_prof,
  flight_speed,
  body_lt,
  wing_span,
  prop_upwind = 0.5,
  flap_glide,
  rotor_speed,
  rotor_radius,
  blade_width,
  blade_pitch,
  n_blades
)
}
\arguments{
\item{chord_prof}{a data.frame with the blade's chord taper profile. Function
expects two named columns:
\itemize{
 \item{\code{pp_radius}, radius at bird passage point, as a proportion
    of \code{rotor_radius}}
 \item{\code{chord}, chord width at (\code{pp_radius}), as a proportion
    of \code{blade_width}}
}}

\item{flight_speed}{The bird flying speed (\eqn{v}), in metres/sec}

\item{body_lt}{The length of the bird (\eqn{L}), in metres}

\item{wing_span}{The wingspan of the bird (\eqn{W}), in metres}

\item{prop_upwind}{A value between 0-1 bounded as proportion of flights
upwind - default of 0.5.}

\item{flap_glide}{A decimal value representing the correction for flapping
or gliding birds (\eqn{F}).}

\item{rotor_speed}{operational rotation speed, in revolutions/min}

\item{rotor_radius}{The radius of the rotor (\eqn{R}), in metres}

\item{blade_width}{The maximum blade width, in metres}

\item{blade_pitch}{The average blade pitch angle, the angle between the blade surface and
the rotor plane (\eqn{\gamma}), in radians}

\item{n_blades}{The number of blades in rotor (\eqn{b}).}
}
\value{
A numeric value. The average collision probability (risk) for a bird
  flying through a rotor's circle area
}
\description{
Calculate the average probability of collision for a single bird rotor transit
under the basic model (i.e. uniform distribution of bird flights at
risk height), assuming no avoidance action.
}
\details{
Methodology and assumptions underpinning \code{get_prob_collision}
are described in "Stage C" of
\href{https://www.bto.org/sites/default/files/u28/downloads/Projects/Final_Report_SOSS02_Band1ModelGuidance.pdf}{Band (2012)}
}
