% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rotor_transits.R
\name{get_rotor_transits}
\alias{get_rotor_transits}
\title{Number of birds transiting through rotors assuming no avoidance}
\usage{
get_rotor_transits(flux_factor, prop_crh)
}
\arguments{
\item{flux_factor}{a vector containing the flux factor for each month}

\item{prop_crh}{The proportion of flights at risk height (\eqn{Q_2R} or
(\eqn{Q'_2R})}
}
\value{
A vector with the number of bird flights potentially transiting through rotors
at each time period (assuming no avoidance)
}
\description{
Calculate the potential number of bird flights transiting through the rotors
of the wind farm per month, assuming birds take no avoiding action
}
\details{
Methodology and assumptions underpinning \code{get_rotor_transits}
are described in "Stage B" of
\href{https://www.bto.org/sites/default/files/u28/downloads/Projects/Final_Report_SOSS02_Band1ModelGuidance.pdf}{Band (2012)}
}
\seealso{
\code{\link[=get_flux_factor]{get_flux_factor()}} for calculating the flux factor
}
