% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleProp.R
\name{GetSampleProp}
\alias{GetSampleProp}
\title{Sample lines in a windfarm area}
\usage{
GetSampleProp(maskedlines, samplesize, WFarea)
}
\arguments{
\item{maskedlines}{An sf object. The full set of lines sampled within the migration corridor for a species}

\item{samplesize}{An integer. The number of lines to sample from the full set of lines}

\item{WFarea}{An sf object. The polygon of the wind farm area of interest}
}
\value{
The number of lines that overlap with the polygons of interest
}
\description{
A function to sample n number of lines from a full set of polylines and get the proportion of those linee
that intersect the windfarm. Used to calculate the population size in migration CRM
}
