% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_flux.R
\name{initial_flux}
\alias{initial_flux}
\title{The initial flux of birds}
\usage{
initial_flux(
  TurbineData = TurbineData,
  sampledTurbine = sampledTurbine,
  sampledBirdParams = sampledBirdParams,
  sampledSpeciesCount = sampledSpeciesCount,
  TPower = TPower,
  hours = hours
)
}
\arguments{
\item{TurbineData}{A data frame. The Turbine data formatted as per the "TurbineData" data object}

\item{sampledTurbine}{A data frame. The data frame with all the sampled parameters from the turbine}

\item{sampledBirdParams}{A data frame. The data frame with all the sampled parameters of the bird's features}

\item{sampledSpeciesCount}{A data frame. The sample of the density of birds in the wind farm area (i.e. the count)}

\item{TPower}{An integer. The total approximate power meant to be generated by the wind farm (TurbineModel field)}

\item{hours}{A data frame. The total number of hours per day at the latitude calculated by DayLength()}
}
\value{
A data frame. The hours data frame returned with the flux calculations included
}
\description{
The flux of birds travelling through the turbine area
}
