% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{check if a number is whole}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric value}

\item{tol}{a numeric value}
}
\description{
check if a number is whole
}
