% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Johnston_flight_heights}
\alias{Johnston_flight_heights}
\title{Bird parameters for the CRM}
\format{
A dataframe
\describe{
\item{Species}{Species name}
\item{AvoidanceBasic}{Basic avoidance rate}
\item{AvoidanceBasicSD}{Basic avoidance rate SD}
...
}
}
\usage{
Johnston_flight_heights
}
\description{
A dataframe of species biological parameters for the model
}
\keyword{datasets}
