% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_array_correction.R
\name{large_array_correction}
\alias{large_array_correction}
\title{Large array correction}
\usage{
large_array_correction(
  NTurbines = NTurbines,
  sampledTurbine = sampledTurbine,
  sampledBirdParams = sampledBirdParams,
  P_Collision = P_Collision,
  MeanOperational = MeanOperational,
  WFWidth = WFWidth
)
}
\arguments{
\item{NTurbines}{An integer value. The number of turbines in the site.}

\item{sampledTurbine}{A data frame. The data frame with all the sampled parameters from the turbine}

\item{sampledBirdParams}{A data frame. The data frame with all the sampled parameters of the bird's features}

\item{P_Collision}{A numeric value. The proportion of birds that could collide the turbine assuming no avoidance}

\item{MeanOperational}{A numeric value. The mean operational time of the wind farm (in hours)}

\item{WFWidth}{A numeric value. The approximate longitudinal width (in km) of the wind farm}
}
\value{
A numeric value. The correction factor to be applied
}
\description{
A correction factor in the case that there is a very large wind farm array which may deter birds
}
