% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_stoch_crm.r
\name{mig_stoch_crm}
\alias{mig_stoch_crm}
\title{Stochastic migration collision risk model for a single species and one turbine scenario}
\usage{
mig_stoch_crm(
  BirdData,
  TurbineData,
  wf_latitude,
  CountData,
  chord_profile = chord_prof_5MW,
  iter = 10,
  spp_name = "",
  LargeArrayCorrection = TRUE
)
}
\arguments{
\item{BirdData}{A data frame. One row of the BirdData table}

\item{TurbineData}{A data frame. One row of the TurbineData field}

\item{CountData}{A data frame. One row of the Count Data Field.}

\item{chord_profile}{A data frame with the chord taper profile of the rotor
blade. It must contain the columns:
\itemize{
\item \code{pp_radius}, equidistant intervals of radius at bird passage point,
as a proportion of \code{rotor_radius}, within the range \eqn{[0, 1]}.
\item \code{chord}, the chord width at \code{pp_radius}, as a proportion of \code{blade_width}.
}

Defaults to a generic profile for a typical modern 5MW turbine. See
\code{\link[=chord_prof_5MW]{chord_prof_5MW()}} for details.}

\item{iter}{An integer constant > 0. The number of stochastic draws to take}

\item{spp_name}{A character vector.}

\item{n_turbines}{integer}
}
\value{
Estimates of number of collisions per migratory season for the number of iters specified
}
\description{
Run migration stochastic collision risk model for a single species and one turbine scenario
}
\details{
This function is an adaption of code from Masden(2015) used for estimating
the collision risk of seabirds in offshore windfarm sites and is a further
adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where
flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function \code{stoch_crm}
\itemize{
\item function will run only option 1 for migratory species
}
}
