% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parent_fn_name}
\alias{parent_fn_name}
\title{Name of parent function}
\usage{
parent_fn_name()
}
\value{
The name of the parent function relative to current evaluation. Returns "GlobalEnv"
if parent frame is the Global environment.
}
\description{
Helper to get the name of the parent function.
}
