% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrisksum2.R
\name{pcoll}
\alias{pcoll}
\title{prob of collision, called within xrisksum2 and pcollxy}
\usage{
pcoll(
  r,
  phi,
  direction,
  rotor_radius,
  blade_width,
  rotor_speed,
  blade_pitch,
  flight_type_num,
  wing_span,
  flight_speed,
  body_lt,
  n_blades
)
}
\arguments{
\item{flight_type}{a numeric value. 1 = flapping, 0 = gliding, indicating the
type of flight that best typifies most flights of the species in question}
}
\description{
prob of collision, called within xrisksum2 and pcollxy
}
