% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrisksum2.R
\name{pcollxy}
\alias{pcollxy}
\title{In Masden construction, relies on various objects from outside function
These will be passed explicitly:
sampledTurbine \verb{[ith row]}
sampledBirdParams \verb{[ith row]}}
\usage{
pcollxy(x, y, direction, pars)
}
\arguments{
\item{x, y, direction, pars}{numeric}
}
\description{
In Masden construction, relies on various objects from outside function
These will be passed explicitly:
sampledTurbine \verb{[ith row]}
sampledBirdParams \verb{[ith row]}
}
