% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_dmp.R
\name{rbeta_dmp}
\alias{rbeta_dmp}
\title{Customised sampling functions for the Beta distributions}
\usage{
rbeta_dmp(n, p, sd)
}
\arguments{
\item{n}{An integer value. The number of samples to generate}

\item{p}{A decimal value. The value used to calculate parameters for the beta distribution}

\item{sd}{A decimal value. The standard deviation of the beta distribution to simulate}
}
\value{
a vector of samples values from the beta distribution
}
\description{
generate random samples from a beta distribution, parameterized as mean and sd, and returning NAs if conditions are not met
}
\examples{
    rbeta_dmp(n=100,p=0.14,sd=9.8)
}
