% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotor_grids_test}
\alias{rotor_grids_test}
\title{Sample rotor grids for generated_rotor_grids unit test}
\format{
list of data frames
}
\usage{
rotor_grids_test
}
\description{
Sample rotor grids for generated_rotor_grids unit test
}
\keyword{datasets}
