% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtnorm_dmp.R
\name{rtnorm_dmp}
\alias{rtnorm_dmp}
\title{Customised sampling functions for the Truncated Normal distribution}
\usage{
rtnorm_dmp(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{An integer value. The number of samples to generate}

\item{mean}{A decimal value. The mean for the truncated normal distribution}

\item{sd}{A decimal value. The standard deviation of the distribution to simulate}

\item{lower}{A decimal value. The lower limit for the distribution}

\item{upper}{A decimal value. The upper limit for the distribution}
}
\value{
a vector of samples values from the truncated normal distribution
}
\description{
generate random samples from a truncated normal distribution, parameterized as mean and sd, and returning NAs if conditions are not met
}
