% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_turbine.R
\name{sample_turbine}
\alias{sample_turbine}
\title{Sampling function for a single turbine}
\usage{
sample_turbine(
  TurbineData = TurbineData,
  windSpeedMean = windSpeedMean,
  windSpeedSD = windSpeedSD,
  windData = windData,
  windThreshold = windThreshold,
  iter = iter
)
}
\arguments{
\item{TurbineData}{A data frame. The Turbine data formatted as per the "TurbineData" data object}

\item{windSpeedMean}{A numeric value. The mean wind speed in km/hr}

\item{windSpeedSD}{A numeric value. The SD of the wind speed in km/hr}

\item{windData}{A data frame.  The wind speed x rotation speed x blade pitch data table}

\item{windThreshold}{A numeric value. The minimum wind speed required for the blades to turn}

\item{iter}{An integer value. The number of samples to generate}
}
\value{
A data frame of all the information sampled for the turbine with nrow = iter
}
\description{
Samples and aggregates appropriate data for a single wind turbine
}
