% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleCount_pctiles}
\alias{sampleCount_pctiles}
\title{Sample count percentiles}
\usage{
sampleCount_pctiles(n, probs, countsPctls)
}
\value{
interpolated random probabilities as a vector
}
\description{
resample from empirical cdf based on set of quantiles
}
