% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCRM_option1.R
\name{sCRM_option1}
\alias{sCRM_option1}
\title{Option 1 of the stochastic Band model}
\usage{
sCRM_option1(
  MonthlyOperational = MonthlyOperational,
  hours = hours,
  PCH = PCH,
  P_Collision = P_Collision,
  AvoidanceBasic = AvoidanceBasic,
  LargeArrayCorrection = FALSE,
  L_ArrayCF = NULL
)
}
\arguments{
\item{MonthlyOperational}{An integer value. The number of turbines in the site.}

\item{hours}{A data frame. The total number of hours per day at the latitude calculated by DayLength()}

\item{PCH}{A numeric value. The percentage of birds at collision risk height}

\item{P_Collision}{A numeric value. The proportion of birds that could collide the turbine assuming no avoidance}

\item{AvoidanceBasic}{A numeric value. The basic avoidance rate as defined by Band (2012) and Masden (2015)}

\item{LargeArrayCorrection}{A boolean. If true, the large array correction will be applied}

\item{L_ArrayCF}{A numeric value. The large array correction factor}
}
\value{
A data frame. The collision rates for each time period.
}
\description{
This runs option 1 of the sCRM as per Masden(2015)
}
