% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCRM_option2.R
\name{sCRM_option2}
\alias{sCRM_option2}
\title{Option 2 of the stochastic Band model}
\usage{
sCRM_option2(
  FlightHeightSpec = FlightHeightSpec,
  MonthlyOperational = MonthlyOperational,
  hours = hours,
  P_Collision = P_Collision,
  sampledTurbine = sampledTurbine,
  TideOff = TideOff,
  sampledBirdParams = sampledBirdParams,
  LargeArrayCorrection = FALSE,
  L_ArrayCF = NULL
)
}
\arguments{
\item{FlightHeightSpec}{A dataframe. The bootstrapped flight height distribution of the species by Johnston et al. 2014}

\item{MonthlyOperational}{An integer value. The number of turbines in the site.}

\item{hours}{A data frame. The total number of hours per day at the latitude calculated by DayLength()}

\item{P_Collision}{A numeric value. The proportion of birds that could collide the turbine assuming no avoidance}

\item{sampledTurbine}{A data frame. The row of the sampled Turbine dataframe that contains turbine information}

\item{TideOff}{A numeric value. The tidal offset}

\item{sampledBirdParams}{A data frame. The row of the sampled bird parameters dataframe that contains PCH and Avoidance}

\item{LargeArrayCorrection}{A boolean. If true, the large array correction will be applied}

\item{L_ArrayCF}{A numeric value. The large array correction factor}
}
\value{
A data frame. The collision rates for each time period.
}
\description{
This runs option 2 of the sCRM as per Masden(2015) - collisions as returned from a smooth distribution (Johnston et al default)
}
