% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_months}
\alias{seq_months}
\title{Generate sequence of months}
\usage{
seq_months(start_month, end_month)
}
\arguments{
\item{start_month}{character string, the three-letter name of the starting month.}

\item{end_month}{character string, the three-letter name of the finishing month.}
}
\value{
character vector. The list of months that falls in between two months
}
\description{
Generate sequence of months
}
\examples{
   seq_months("Jan", "Apr")
}
