% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{startUpValues}
\alias{startUpValues}
\title{Start up values for the mCRM}
\format{
A nested list with 3 upper levels
\describe{
\item{turbinepars}{Parameters for the wind farm turbines}
\item{windfarmpars}{Parameters for the wind farms themselves}
\item{speciespars}{Parameters for species}
...
}
}
\usage{
startUpValues
}
\description{
A dataset of startup values for the app to run
}
\keyword{datasets}
