% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoch_crm.r
\name{stoch_crm}
\alias{stoch_crm}
\title{Stochastic collision risk model for a single species and one turbine scenario}
\usage{
stoch_crm(
  model_options = c(1, 2, 3),
  BirdData,
  TurbineData,
  CountData,
  iter = 10,
  spp_name,
  LargeArrayCorrection,
  n_turbines,
  WFWidth,
  Prop_Upwind,
  Latitude,
  TideOff,
  windSpeedMean,
  windSpeedSD,
  windData_rotation,
  windData_pitch,
  fhd_bootstraps = NULL,
  dens_opt = "truncNorm"
)
}
\arguments{
\item{model_options}{numeric vector}

\item{BirdData}{A data frame. Contains all the parameters for the species}

\item{TurbineData}{A data frame. Contains all the parameters for the turbines}

\item{CountData}{A data frame. Contains density data for the wind farm of interest}

\item{iter}{An integer constant > 0. The number of stochastic draws to take}

\item{spp_name}{A character vector.}

\item{LargeArrayCorrection}{A boolean. If TRUE, correct if the wind farm array is very large}

\item{n_turbines}{integer}

\item{WFWidth}{A decimal value. The "width" of wind farm used in Large Array Correction (KM)}

\item{Prop_Upwind}{A decimal value. A value between 0-1 bounded as proportion of flights upwind - default of 0.5.}

\item{TideOff}{A decimal value. Tidal offset in metres}

\item{windSpeedMean}{A decimal value. Site specific mean wind speed (m/s)}

\item{windSpeedSD}{A decimal value. Site specific standard deviation of wind speeds}

\item{windData_rotation}{A data frame. The table of wind speed versus rotor speed}

\item{windData_pitch}{A data frame. The table of wind speed versus rotor pitch}

\item{dens_opt}{A character value. The type of sampling to do for the species density data.}

\item{FlightData}{A data frame. Contains flight height data for species}

\item{TPower}{A decimal value. The amount of power generated by the wind farm (MW)}

\item{Latitude.}{A decimal value. Latitude in WGS 1984 (decimal degrees)}
}
\value{
Estimates of number of collisions per month, for each of the chosen
   model options
}
\description{
Run stochastic collision risk model for a single species and one turbine scenario
}
\details{
This function is an adaption of code from Masden(2015) used for estimating
the collision risk of seabirds in offshore windfarm sites and is a further
adaptation from Band(2012).

The collision risk model evaluates risk on a month by month basis in order
to reflect changing bird abundance within and utilization of the area.

Changes in relation to previous top-line function \code{stochasticBand}
\itemize{
  \item function runs the model for one single scenario (i.e. one species for
  one turbine scenario). Advantages include:
  \itemize{
      \item streamlined infrastructure for easier scenario management
      \item easier implementation of parallelisation for multiple scenarios
  }
  \item results are no longer saved to an external file
}
}
\examples{
stoch_crm(model_options = c(1, 2, 3),
  BirdData = Bird_Data[3, ],
  TurbineData = Turbine_Data,
  CountData = Count_Data[3, ],
  iter = 10,
  spp_name = c("Black_legged_Kittiwake"),
  LargeArrayCorrection = TRUE,
  n_turbines = 300,
  WFWidth = 4,
  Prop_Upwind = 0.5,
  Latitude = 56,
  TideOff = 2.5,
  windSpeedMean = 30,
  windSpeedSD = 5.1,
  windData_rotation = startUpValues$turbinePars$rotationVsWind_df,
  windData_pitch = startUpValues$turbinePars$pitchVsWind_df,
  dens_opt = "truncNorm",
  fhd_bootstraps = generic_fhd_bootstraps$Black_headed_Gull)

}
