% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_crm_option1.R
\name{stochasticBand}
\alias{stochasticBand}
\title{Translation of Masden 2015 collision risk models for offshore wind collisions}
\usage{
stochasticBand(
  results_folder = NULL,
  BirdDataFile = "data/BirdData.csv",
  TurbineDataFile = "data/TurbineData.csv",
  CountDataFile = "data/CountData.csv",
  FlightDataFile = "data/FlightHeight.csv",
  iter = 10,
  CRSpecies = c("Northern_Gannet"),
  TPower = 1760,
  LargeArrayCorrection = "yes",
  WFWidth = 54,
  Prop_Upwind = 0.5,
  Latitude = 53.7,
  TideOff = 2.5,
  windSpeedMean = 30,
  windSpeedSD = 1e-09,
  DensityOpt
)
}
\arguments{
\item{results_folder}{A character string. Path name to an existing output folder for tables/plots}

\item{BirdDataFile}{A data frame. Contains all the parameters for the species}

\item{TurbineDataFile}{A data frame. Contains all the parameters for the turbines}

\item{CountDataFile}{A data frame. Contains density data for the wind farm of interest}

\item{FlightDataFile}{A data frame. Contains flight height data for species}

\item{iter}{An integer constant > 0. The number of stochastic draws to take}

\item{CRSpecies}{A character vector. List of species to consider. Looks to match/define other species name, so must match throughout}

\item{TPower}{A decimal value. The amount of power generated by the wind farm (MW)}

\item{LargeArrayCorrection}{A character vector. Correct for a large array? "yes" or "no"}

\item{WFWidth}{A decimal value. The "width" of wind farm used in Large Array Correction (KM)}

\item{Prop_Upwind}{A decimal value. A value between 0-1 bounded as proportion of flights upwind - default of 0.5.}

\item{TideOff}{A decimal value. Tidal offset in metres}

\item{windSpeedMean}{A decimal value. Site specific mean wind speed (m/s)}

\item{windSpeedSD}{A decimal value. Site specific standard deviation of wind speeds}

\item{Latitude.}{A decimal value. Latitude in WGS 1984 (decimal degrees)}
}
\value{
CSV files saved into the results_folder directory
}
\description{
This function is an adaption of code from Masden(2015) used for estimating the collision risk of seabirds in
offshore windfarm sites and is a further adaptation from Band(2012).
}
