% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Turbine_Data}
\alias{Turbine_Data}
\title{Bird parameters for the CRM}
\format{
A dataframe
\describe{
  \item{Species}{Species name}
  \item{AvoidanceBasic}{Basic avoidance rate}
  \item{AvoidanceBasicSD}{Basic avoidance rate SD}
  ...
}
}
\usage{
Turbine_Data
}
\description{
A dataframe of species biological parameters for the model
}
\keyword{datasets}
