% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wndspd_rtn_ptch_example}
\alias{wndspd_rtn_ptch_example}
\title{Example of data with relationship between wind speed, rotation speed and
blade pitch}
\format{
A 30 x 3 data frame with columns:
\describe{
\item{wind_speed}{Wind speed in metres per second.}
\item{rtn_speed}{Blade rotation speed, in revolutions per minute}
\item{bld_pitch}{Blade pitch, in degrees}
}
}
\usage{
wndspd_rtn_ptch_example
}
\description{
Intended to illustrate the application of \code{stoch_scrm()}
}
\keyword{datasets}
