% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrisksum2.R
\name{xrisksum2}
\alias{xrisksum2}
\title{The xrisksum2 function for calculating risk along a turbine blade}
\usage{
xrisksum2(
  y,
  xinc,
  direction,
  inputRad = currentRad,
  inputCirc = currentCirc,
  inputRotorRadius = currentRotorRadius,
  inputBladeWidth = currentBladeWidth,
  inputRotorSpeed = currentRotorSpeed,
  inputPitch = currentPitch,
  inputFlight = currentFlightNumeric,
  inputWingSpan = currentWingSpan,
  inputFlightSpeed = currentFlightSpeed,
  inputBirdLength = currentBirdLength,
  inputBlades = currentBlades
)
}
\description{
This function will calculate the risk associated with each point along the surface of a turbine blade
}
