% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_opt1.R
\name{crm_opt1}
\alias{crm_opt1}
\title{Number of collisions under model option 1}
\usage{
crm_opt1(
  flux_factor,
  prop_crh_surv,
  prob_single_collision,
  prop_operational,
  avoidance_rate,
  lac_factor
)
}
\arguments{
\item{flux_factor}{a vector containing the flux factor for each month}

\item{prop_crh_surv}{The proportion of flights at collision risk height derived
from site survey (\eqn{Q_2R})}

\item{prob_single_collision}{The probability of collision for a single bird
transit through a rotor, assuming no avoidance action.}

\item{prop_operational}{The proportion of time during which turbines are
operational per month}

\item{avoidance_rate}{The proportion of birds likely to take action to avoid
collision with rotor}

\item{lac_factor}{The large array correction factor. Defaults to 1,
meaning large array correction is not applicable}
}
\value{
Expected number of collisions per month based on model option 1
}
\description{
Top-level function to run CRM model under option 1:
   \itemize{
      \item Basic model, i.e. flights across collision risk height are
       uniformly distributed
      \item proportion at collision risk height derived from site survey
   }
}
