% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DayLength.r
\name{DayLength}
\alias{DayLength}
\title{Day length function}
\usage{
DayLength(Latitude)
}
\arguments{
\item{Latitude}{A decimal value. The latitude of the centroid of the windfarm}
}
\value{
data frame of hours
}
\description{
Taken from Forsythe et al.(1995) A model comparison for daylength as a
function of latitude and day of year.  Ecological Modelling. 80: 87 - 95
}
