% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_months}
\alias{format_months}
\title{Format any month name to three letter code}
\usage{
format_months(months)
}
\arguments{
\item{months}{character string or vector. The name of the month}
}
\description{
Format any month name to three letter code
}
\examples{
  format_months("January")
}
