% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_y.R
\name{get_risk_y}
\alias{get_risk_y}
\title{Single transit collision risk along the chord of the rotor at height band
\eqn{y}}
\usage{
get_risk_y(x_at_y, pcoll_doty)
}
\arguments{
\item{x_at_y}{a vector, sequence of horizontal distances from rotor's
vertical axis to points \eqn{x} along half of the rotor circle, expressed
as the proportion of rotor radius, at height band \eqn{y}}

\item{pcoll_doty}{a vector, the single transit collision risk at horizontal
distances \eqn{x}, at height band \eqn{y}}
}
\value{
a numeric value, the single transit collision risk along the whole
horizontal chord of the rotor circle at height band \eqn{y}
}
\description{
Calculates the single transit collision risk/probability along the horizontal
chord of the rotor at height \eqn{y} via numerical integration.
}
\seealso{
\code{\link[=get_x_grid]{get_x_grid()}}, \code{\link[=get_pcoll_grid]{get_pcoll_grid()}}
}
