% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mig_stoch_crm.r
\name{mig_stoch_crm}
\alias{mig_stoch_crm}
\title{Stochastic migration collision risk model for a single species and one turbine scenario}
\usage{
mig_stoch_crm(
  BirdData,
  TurbineData,
  CountData,
  iter = 10,
  spp_name = "",
  LargeArrayCorrection = TRUE
)
}
\arguments{
\item{BirdData}{A data frame. One row of the BirdData table}

\item{TurbineData}{A data frame. One row of the TurbineData field}

\item{CountData}{A data frame. One row of the Count Data Field.}

\item{iter}{An integer constant > 0. The number of stochastic draws to take}

\item{spp_name}{A character vector.}

\item{n_turbines}{integer}
}
\value{
Estimates of number of collisions per migratory season for the number of iters specified
}
\description{
Run migration stochastic collision risk model for a single species and one turbine scenario
}
\details{
This function is an adaption of code from Masden(2015) used for estimating
the collision risk of seabirds in offshore windfarm sites and is a further
adaptation from Band(2012). It is a further adaptation of the stoch_crm function.

The collision risk model evaluates risk for each defined migratory period where
flux rate is simply the number of birds travelling through the windfarm.

Changes in relation to previous top-line function \code{stoch_crm}
\itemize{
\item function will run only option 1 for migratory species
}
}
