% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrisksum2.R
\name{pcollxy}
\alias{pcollxy}
\title{In Masden construction, relies on various objects from outside function
These will be passed explicitly:
sampledTurbine [ith row]
sampledBirdParams [ith row]}
\usage{
pcollxy(x, y, direction, pars)
}
\description{
In Masden construction, relies on various objects from outside function
These will be passed explicitly:
sampledTurbine [ith row]
sampledBirdParams [ith row]
}
