% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_turbine_mCRM.R
\name{sample_turbine_mCRM}
\alias{sample_turbine_mCRM}
\title{Sampling function for a single turbine in the mCRM}
\usage{
sample_turbine_mCRM(TurbineData, BirdData, iter)
}
\arguments{
\item{TurbineData}{A data frame. The Turbine data formatted as per the "TurbineData" data object}

\item{BirdData}{A data frame. The bird data for getting migratory seasons}

\item{iter}{An integer value. The number of samples to generate}
}
\value{
A data frame of all the information sampled for the turbine with nrow = iter
}
\description{
Samples and aggregates appropriate data for a single wind turbine
}
